Split Comics into container and presentational components for reusability
